Get tests passing again
