--- packages/gatsby-remark-shiki/src/utils.d.ts
@@ -1,3 +1,0 @@
-type Range = { begin: number; end: number; text?: string; count?: number; tooltip?: string[] }
-
-export function createHighlightedString(ranges: Range[], text: string): string
